==> differential_analysis/tissue_comparison/celltype_enrichment/song_ming/_h/03.plot_heatmap.R <==
library(ggplot2)
library(tidyverse)

save_plot <- function(p, fn, w, h){
    for(ext in c(".pdf", ".svg")){
        ggsave(filename=paste0(fn,ext), plot=p, width=w, height=h)
    }
}

load_celltype_enrichment <- function(dataset){
    fn <- paste0("celltype_enrichment_analysis.",
                 tolower(dataset), ".txt")
    return(data.table::fread(fn))
}
memENRICH <- memoise::memoise(load_celltype_enrichment)

gen_data <- function(dataset){
    err = 0.0000001
    dt <- memENRICH(dataset) |> mutate_if(is.character, as.factor) |>
        mutate(Cell_type=fct_relevel(Cell_type, rev), `-log10(FDR)`= -log10(FDR),
               `OR Percentile`= OR / (1+OR), p.fdr.sig=FDR < 0.05,
               `log2(OR)` = log2(OR+err),
               p.fdr.cat=cut(FDR, breaks=c(1,0.05,0.01,0.005,0),
                             labels=c("<= 0.005","<= 0.01","<= 0.05","> 0.05"),
                             include.lowest=TRUE)) |>
        mutate(across(Direction, \(x) factor(x, levels=c("All", "Down", "Up"))))
    levels(dt$Direction) <- c("All", "AA Bias", "EA Bias")
    return(dt)
}
memDF <- memoise::memoise(gen_data)

plot_tile <- function(dataset, w, h){
    y0 <- min(memDF(dataset)$`log2(OR)`)-0.1
    y1 <- max(memDF(dataset)$`log2(OR)`)+0.1
    tile_plot <- memDF(dataset) |>
        ggplot(aes(y = Cell_type, x = Tissue, fill = `log2(OR)`,
                   label = ifelse(p.fdr.sig,
                                  format(round(`-log10(FDR)`,1), nsmall=1), ""))) +
        xlab('Ancestry-Related DEGs') + ylab("Celltypes") +
        facet_grid(.~Direction) + geom_tile(color = "grey") +
        ggfittext::geom_fit_text(contrast = TRUE) +
        scale_fill_gradientn(colors=c("blue", "white", "red"),
                             values=scales::rescale(c(y0, 0, y1)),
                             limits=c(y0,y1)) +
        ggpubr::theme_pubr(base_size = 20, border=FALSE) +
        theme(axis.text.x = element_text(angle = 45, hjust=1),
              legend.position="right",
              axis.title=element_text(face="bold", size=28),
              strip.text=element_text(face="bold", size=22))
    outfile <- paste0("tileplot_enrichment.",tolower(dataset))
    save_plot(tile_plot, outfile, w, h)
}

## Run script
plot_tile("microglia", 12, 7)
plot_tile("astrocyte", 12, 8)
plot_tile("oligo", 12, 7)

## Reproducibility information
Sys.time()
proc.time()
options(width = 120)
sessioninfo::session_info()
